+++ internal/controller/monovertexrollout/monovertexrollout_progressive.go
@@ -8,9 +8,99 @@ import (
 	"github.com/numaproj/numaplane/internal/util/logger"
 	apiv1 "github.com/numaproj/numaplane/pkg/apis/numaplane/v1alpha1"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
+
+	"sigs.k8s.io/controller-runtime/pkg/client"
+
+	"github.com/numaproj/numaplane/internal/common"
+	ctlrcommon "github.com/numaproj/numaplane/internal/controller/common"
+	"github.com/numaproj/numaplane/internal/controller/common/numaflowtypes"
+	"github.com/numaproj/numaplane/internal/util"
 )
 
-// TODO: can move other progressive interface functions here as well? (same with PipelineRollout, ISBServiceRollout)
+// Implemented functions for the progressiveController interface:
+
+func (r *MonoVertexRolloutReconciler) CreateUpgradingChildDefinition(ctx context.Context, rolloutObject ctlrcommon.RolloutObject, name string) (*unstructured.Unstructured, error) {
+	monoVertexRollout := rolloutObject.(*apiv1.MonoVertexRollout)
+	metadata, err := getBaseMonoVertexMetadata(monoVertexRollout)
+	if err != nil {
+		return nil, err
+	}
+	monoVertex, err := r.makeMonoVertexDefinition(monoVertexRollout, name, metadata)
+	if err != nil {
+		return nil, err
+	}
+
+	labels := monoVertex.GetLabels()
+	labels[common.LabelKeyUpgradeState] = string(common.LabelValueUpgradeInProgress)
+	monoVertex.SetLabels(labels)
+
+	return monoVertex, nil
+}
+
+func (r *MonoVertexRolloutReconciler) getCurrentChildCount(rolloutObject ctlrcommon.RolloutObject) (int32, bool) {
+	monoVertexRollout := rolloutObject.(*apiv1.MonoVertexRollout)
+	if monoVertexRollout.Status.NameCount == nil {
+		return int32(0), false
+	} else {
+		return *monoVertexRollout.Status.NameCount, true
+	}
+}
+
+func (r *MonoVertexRolloutReconciler) updateCurrentChildCount(ctx context.Context, rolloutObject ctlrcommon.RolloutObject, nameCount int32) error {
+	monoVertexRollout := rolloutObject.(*apiv1.MonoVertexRollout)
+	monoVertexRollout.Status.NameCount = &nameCount
+	return r.updateMonoVertexRolloutStatus(ctx, monoVertexRollout)
+}
+
+// increment the child count for the Rollout and return the count to use
+func (r *MonoVertexRolloutReconciler) IncrementChildCount(ctx context.Context, rolloutObject ctlrcommon.RolloutObject) (int32, error) {
+	currentNameCount, found := r.getCurrentChildCount(rolloutObject)
+	if !found {
+		currentNameCount = int32(0)
+		err := r.updateCurrentChildCount(ctx, rolloutObject, int32(0))
+		if err != nil {
+			return int32(0), err
+		}
+	}
+
+	err := r.updateCurrentChildCount(ctx, rolloutObject, currentNameCount+1)
+	if err != nil {
+		return int32(0), err
+	}
+	return currentNameCount, nil
+}
+
+func (r *MonoVertexRolloutReconciler) Recycle(ctx context.Context,
+	monoVertexDef *unstructured.Unstructured,
+	c client.Client,
+) error {
+	return kubernetes.DeleteResource(ctx, c, monoVertexDef)
+}
+
+// ChildNeedsUpdating() tests for essential equality, with any irrelevant fields eliminated from the comparison
+func (r *MonoVertexRolloutReconciler) ChildNeedsUpdating(ctx context.Context, from, to *unstructured.Unstructured) (bool, error) {
+	numaLogger := logger.FromContext(ctx)
+	// remove "replicas" field from comparison to test for equality
+	fromNew, err := numaflowtypes.MonoVertexWithoutReplicas(from)
+	if err != nil {
+		return false, err
+	}
+	toNew, err := numaflowtypes.MonoVertexWithoutReplicas(to)
+	if err != nil {
+		return false, err
+	}
+
+	specsEqual := util.CompareStructNumTypeAgnostic(fromNew, toNew)
+	numaLogger.Debugf("specsEqual: %t, fromNew=%v, toNew=%v\n",
+		specsEqual, fromNew, toNew)
+	labelsEqual := util.CompareMaps(from.GetLabels(), to.GetLabels())
+	numaLogger.Debugf("labelsEqual: %t, from Labels=%v, to Labels=%v", labelsEqual, from.GetLabels(), to.GetLabels())
+	annotationsEqual := util.CompareMaps(from.GetAnnotations(), to.GetAnnotations())
+	numaLogger.Debugf("annotationsEqual: %t, from Annotations=%v, to Annotations=%v", annotationsEqual, from.GetAnnotations(), to.GetAnnotations())
+
+	return !specsEqual || !labelsEqual || !annotationsEqual, nil
+
+}
 
 // make an assessment of the upgrading child to determine if it was successful, failed, or still not known
 // TODO: fix this assessment not to return an immediate result as soon as things are healthy or unhealthy
